[Messenger] added interfaces to be type-hinted even when not using a Container
