+++ src/Symfony/Component/Messenger/Handler/Locator/HandlerLocator.php
@@ -11,12 +11,10 @@
 
 namespace Symfony\Component\Messenger\Handler\Locator;
 
-use Symfony\Component\Messenger\Exception\NoHandlerForMessageException;
-
 /**
  * @author Samuel Roze <samuel.roze@gmail.com>
  */
-class HandlerLocator implements HandlerLocatorInterface
+class HandlerLocator extends AbstractHandlerLocator
 {
     /**
      * Maps a message (its class) to a given handler.
@@ -28,14 +26,8 @@ class HandlerLocator implements HandlerLocatorInterface
         $this->messageToHandlerMapping = $messageToHandlerMapping;
     }
 
-    public function resolve($message): callable
+    protected function getHandler(string $class)
     {
-        $messageKey = \get_class($message);
-
-        if (!isset($this->messageToHandlerMapping[$messageKey])) {
-            throw new NoHandlerForMessageException(sprintf('No handler for message "%s".', $messageKey));
-        }
-
-        return $this->messageToHandlerMapping[$messageKey];
+        return $this->messageToHandlerMapping[$class] ?? null;
     }
 }
